move button and navbar into components folder
