+++ src/js/components/button.js
@@ -1,9 +1,8 @@
 // @flow
 
 import * as React from 'react';
-import '../css/style.scss';
-import { classify } from './tools/tools';
-// import ReactDOM from 'react-dom';
+import '../../css/style.scss';
+import { classify } from '../tools/tools';
 
 type Props = {
   label?: string,
